% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate}
\alias{calculate}
\title{Ultra great function to calculate stock abundance from catch data}
\usage{
calculate(catchdata)
}
\arguments{
\item{catchdata}{Catch data}
}
\description{
Ultra great function to calculate stock abundance from catch data
}
