% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{The metadata of stock assessment}
\format{
A list contains fields below:
\describe{
  \item{species}{species name}
  \item{year}{year of the assessment}
}
}
\usage{
metadata
}
\description{
The metadata of stock assessment
}
\keyword{datasets}
