% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%T>\%}
\alias{\%T>\%}
\title{See \code{magrittr::\link[magrittr:pipe]{\%T>\%}} for details.}
\usage{
lhs \%T>\% rhs
}
\description{
See \code{magrittr::\link[magrittr:pipe]{\%T>\%}} for details.
}
\keyword{internal}
