% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{assess}
\alias{assess}
\title{Ultra great function to calculate stock abundance from catch data}
\usage{
assess(catchdata, coef, variable)
}
\arguments{
\item{catchdata}{Catch data}

\item{coef}{Mulitiplication coefficient (dummy calculation)}

\item{variable}{Variable to return}
}
\description{
Ultra great function to calculate stock abundance from catch data
}
